requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
duckdb
pandas
openai
langchain-groq
duckduckgo_search==5.3.1b1
pymupdf
arxiv
wikipedia
mysql-connector-python
SQLAlchemy
validators==0.28.1
youtube_transcript_api
unstructured
pytube
numexpr
huggingface_hub
langchain_core
langserve
langchain_community
streamlit
